Move return codes to library
