+++ chipsec/modules/common/bios_kbrd_buffer.py
@@ -33,7 +33,8 @@ Examples:
 
 """
 
-from chipsec.module_common import BaseModule, ModuleResult, MTAG_BIOS
+from chipsec.module_common import BaseModule, MTAG_BIOS
+from chipsec.library.returncode import ModuleResult
 from chipsec.logger import print_buffer_bytes
 from typing import List
 
@@ -45,7 +46,8 @@ COMMON_FILL_PTRN = "".join([f'{(chr(x + 0x1E)):1}' for x in range(32)])
 class bios_kbrd_buffer(BaseModule):
     def __init__(self):
         BaseModule.__init__(self)
-        self.rc_res = ModuleResult(0x5ebf705, 'https://chipsec.github.io/modules/chipsec.modules.common.bios_kbrd_buffer.html')
+        self.result.id = 0x5ebf705
+        self.result.url = 'https://chipsec.github.io/modules/chipsec.modules.common.bios_kbrd_buffer.html'
 
     def is_supported(self) -> bool:
         return True
@@ -63,8 +65,8 @@ class bios_kbrd_buffer(BaseModule):
 
         if COMMON_FILL_PTRN == bios_kbrd_buf:
             self.logger.log_good("Keyboard buffer is filled with common fill pattern")
-            self.rc_res.setStatusBit(self.rc_res.status.SUCCESS)
-            return self.rc_res.getReturnCode(ModuleResult.PASSED)
+            self.result.setStatusBit(self.result.status.SUCCESS)
+            return self.result.getReturnCode(ModuleResult.PASSED)
 
         for x in bios_kbrd_buf:
             if ("\x00" != x) and ("\x20" != x):
@@ -85,11 +87,11 @@ class bios_kbrd_buffer(BaseModule):
             self.logger.log_passed("Keyboard buffer looks empty. Pre-boot passwords don't seem to be exposed")
 
         if has_contents:
-            self.rc_res.setStatusBit(self.rc_res.status.POTENTIALLY_VULNERABLE)
-            return self.rc_res.getReturnCode(ModuleResult.WARNING)
+            self.result.setStatusBit(self.result.status.POTENTIALLY_VULNERABLE)
+            return self.result.getReturnCode(ModuleResult.WARNING)
         else:
-            self.rc_res.setStatusBit(self.rc_res.status.SUCCESS)
-            return self.rc_res.getReturnCode(ModuleResult.PASSED)
+            self.result.setStatusBit(self.result.status.SUCCESS)
+            return self.result.getReturnCode(ModuleResult.PASSED)
 
     # --------------------------------------------------------------------------
     # run( module_argv )
